+++ tests/pe/test_forward_information.py
@@ -9,8 +9,12 @@ def test_basic():
     path = get_sample('PE/PE32_x86_library_kernel32.dll')
     sample = lief.parse(path)
     exports = sample.get_export()
-    forwarded_exports = [exp for exp in exports.entries if exp.is_forwarded]
+    forwarded_exports: list[lief.PE.ExportEntry] = [exp for exp in exports.entries if exp.is_forwarded]
     assert len(forwarded_exports) == 82
+
+    assert forwarded_exports[0].name == "InterlockedPushListSList"
+    assert forwarded_exports[0].forward_information.function == "RtlInterlockedPushListSList"
+    assert forwarded_exports[0].forward_information.library == "NTDLL"
 
     print(exports)
     print(exports.entries[0])
@@ -22,4 +26,3 @@ def test_basic():
     assert "forward_information" in json_serialized
     assert json_serialized["forward_information"]["library"] == "NTDLL"
     assert json_serialized["forward_information"]["function"] == "RtlInterlockedPushListSList"
-